+++ packages/host-container/src/types.ts
@@ -89,7 +89,7 @@ export type Container = {
 
   // accounts
 
-  handleAccountGetRoot: InferHandler<'v1', HostApiProtocol['host_account_get_root']>;
+  handleGetUserId: InferHandler<'v1', HostApiProtocol['host_get_user_id']>;
   handleRequestLogin: InferHandler<'v1', HostApiProtocol['host_request_login']>;
   handleAccountConnectionStatusSubscribe: InferHandler<
     'v1',