+++ packages/host-papp-react-ui/.storybook/preview.ts
@@ -1,2 +1,0 @@
-// @ts-expect-error css imports are not defined here
-import '@novasamatech/tr-ui/styles.css';